chore: Refactored visit handler out of cells file
